bennaaym/yaml_support: yaml_support directory refactored
