+++ extensions/yaml_support/labgraph_yaml_parser/_parser/base_parser.py
@@ -3,7 +3,7 @@
 
 from abc import ABCMeta, abstractmethod
 from typing import List
-from extensions.yaml_support.yaml_support.model.base_model import BaseModel
+from extensions.yaml_support.labgraph_yaml_parser.model.base_model import BaseModel
 
 
 class BaseParser(metaclass=ABCMeta):
