+++ extensions/yaml_support/labgraph_yaml_parser/_parser/lg_units_parser.py
@@ -3,8 +3,8 @@
 
 from typed_ast import ast3
 from .base_parser import BaseParser
-from extensions.yaml_support.yaml_support.model.lg_unit_model import LabGraphUnitsModel
-from extensions.yaml_support.yaml_support.enums.lg_units_enum import LabGraphBuiltinUnits
+from extensions.yaml_support.labgraph_yaml_parser.model.lg_unit_model import LabGraphUnitsModel
+from extensions.yaml_support.labgraph_yaml_parser.enums.lg_units_enum import LabGraphBuiltinUnits
 from typed_ast.ast3 import (
     AsyncFunctionDef,
     NodeVisitor,
